src/requirements.txt
boto3
botocore
pyjwt==2.8.0